stream_proxy/streamer/requirements.txt
redis
starlette
fastapi
uvicorn[standard]
PyJWT
python-dotenv
websockets